requirements.txt
Flask==3.0.3
openai==1.27.0
boto3==1.36.26
botocore==1.36.26
Werkzeug==3.0.3
python-dotenv==1.0.1
gunicorn==21.2.0